+++ BudgetPlan.Shared/Dtos/TransactionItemsForDayDto.cs
@@ -2,6 +2,6 @@ namespace BudgetPlan.Shared.Dtos;
 
 public class TransactionItemsForDayDto
 {
-    public float ValuesOfTransactionItemsForDay => TransactionItemDtos.Sum(x => x.Value);
+    public double ValuesOfTransactionItemsForDay => TransactionItemDtos.Sum(x => x.Value);
     public List<TransactionItemDto> TransactionItemDtos { get; set; } = new();
 }
