+++ BudgetPlan.Shared/ViewModels/BudgetPlanViewModel.cs
@@ -5,10 +5,10 @@ namespace BudgetPlan.Shared.ViewModels;
 
 public class BudgetPlanViewModel
 {
-    public float Balance => Income - Expenses;
-    public float Income => BudgetPlanOverTransactionCategoryDtos.Where(x => x.TransactionType == TransactionType.Income)
+    public double Balance => Income - Expenses;
+    public double Income => BudgetPlanOverTransactionCategoryDtos.Where(x => x.TransactionType == TransactionType.Income)
         .Sum(x => x.SumOfAllCategories);
-    public float Expenses => BudgetPlanOverTransactionCategoryDtos.Where(x => x.TransactionType == TransactionType.Expense)
+    public double Expenses => BudgetPlanOverTransactionCategoryDtos.Where(x => x.TransactionType == TransactionType.Expense)
         .Sum(x => x.SumOfAllCategories);
 
     public List<BudgetPlanOverTransactionCategoryDto> BudgetPlanOverTransactionCategoryDtos { get; set; } = new();
